+++ src/yuca/src/main/java/yuca/linux/drive/PowerMode.java
@@ -1,4 +1,4 @@
-package yuca.linux.drive.spec;
+package yuca.linux.drive;
 
 public enum PowerMode {
   STANDBY,
