+++ src/yuca/src/main/java/yuca/linux/drive/commands/BUILD
@@ -1,27 +1,24 @@
-BINARY_SRCS = [""]
-
 java_library(
-    name = "drive",
+    name = "commands",
     visibility = ["//visibility:public"],
-    srcs = glob(["*.java"], exclude = BINARY_SRCS),
+    srcs = glob(["*.java"]),
     deps = [
+        "//src/yuca/src/main/java/yuca/linux/drive:drive",
+        "//src/yuca/src/main/java/yuca/linux/drive/powermode:powermode",
         "//src/yuca/src/main/java/yuca/util",
-        # "//src/yuca/src/main/java/yuca/linux/drive/commands:commands",
         "//src/yuca/src/main/proto/yuca/signal:signal_java_protos",
     ],
-    resources = [
-        "//src/yuca/src/main/c/yuca/drive:libdrivecmds.so",
-        "//src/yuca/src/main/resources/disks:disks_data"
-        ],
 )
 
 java_binary(
     name = "smoke_test",
-    main_class = "yuca.linux.drive.DriveSmokeTest",
+    main_class = "yuca.linux.drive.commands.DriveSmokeTest",
     srcs = ["DriveSmokeTest.java"],
     deps = [
-        ":drive",
+        ":commands",
+        "//src/yuca/src/main/java/yuca/linux/drive/powermode:powermode",
         "//src/yuca/src/main/java/yuca/util",
         "//src/yuca/src/main/proto/yuca/signal:signal_java_protos",
     ],
 )
+
