+++ src/yuca/src/main/java/yuca/linux/drive/powermode/BUILD
@@ -1,8 +1,10 @@
 java_library(
-    name = "spec",
+    name = "powermode",
     visibility = ["//visibility:public"],
     srcs = glob(["*.java"]),
     deps = [
+        "//src/yuca/src/main/java/yuca/linux/drive:drive",
+        "//src/yuca/src/main/java/yuca/util",
         "//src/yuca/src/main/proto/yuca/signal:signal_java_protos",
     ],
 )
